+++ README.md
@@ -3,8 +3,8 @@
 DomainDriven is a batteries included synchronous event sourcing and CQRS library. It is split into the following packages:
 
 - [domaindriven-core](domaindriven-core) - Core persistence model with PostgreSQL and in-memory backends.
-- [domaindriven-effectful](domaindriven-effectful) - Effectful-based API layer with `Aggregate` and `Projection` effects, plus Servant integration.
-- [domaindriven-effectful-examples](domaindriven-effectful-examples) - Example applications demonstrating usage.
+- [domaindriven](domaindriven) - Effectful-based API layer with `Aggregate` and `Projection` effects, plus Servant integration.
+- [domaindriven-examples](domaindriven-examples) - Example applications demonstrating usage.
 
 ## Design idea
