+++ cabal.project
@@ -1,7 +1,7 @@
 packages:
   domaindriven-core/
-  domaindriven-effectful/
-  domaindriven-effectful-examples/
+  domaindriven/
+  domaindriven-examples/
 
 -- Use Stackage LTS 24.2 as the main package source
 import: https://www.stackage.org/lts-24.31/cabal.config
