--- domaindriven-effectful/src/DomainDriven/Effectful.hs
@@ -1,14 +1,0 @@
-module DomainDriven.Effectful
-    ( -- * Domain configuration (re-exported)
-        module DomainDriven.Effectful.Domain
-
-      -- * Effects and helpers
-    , module X
-    )
-where
-
-import DomainDriven.Effectful.Aggregate as X
-import DomainDriven.Effectful.Domain
-import DomainDriven.Effectful.Interpreter as X
-import DomainDriven.Effectful.Projection as X
-import DomainDriven.Persistance.Class as X (Indexed (..), NoIndex (..), Stored (..))
