+++ domaindriven/src/DomainDriven.hs
@@ -1,0 +1,14 @@
+module DomainDriven
+    ( -- * Domain configuration (re-exported)
+        module DomainDriven.Domain
+
+      -- * Effects and helpers
+    , module X
+    )
+where
+
+import DomainDriven.Aggregate as X
+import DomainDriven.Domain
+import DomainDriven.Interpreter as X
+import DomainDriven.Projection as X
+import DomainDriven.Persistance.Class as X (Indexed (..), NoIndex (..), Stored (..))
